@@ -46,12 +46,12 @@
     "stylelint-config-sass-guidelines": "^7.1.0",
     "svg-spritemap-webpack-plugin": ">= 3.7.1",
     "terser-webpack-plugin": "^5.0.3",
-    "webpack": "^5.10.0",
+    "webpack": "^5.76.0",
     "webpack-cli": ">= 3.2.1"
   },
   "dependencies": {
     "@popperjs/core": "^2.6.0",
-    "bootstrap-italia": "^2.0.4",
+    "bootstrap-italia": "^2.8.8",
     "echarts": "^5.4.0",
     "echarts-for-react": "^3.0.2",
     "file-saver": "^2.0.5",
@@ -67,5 +67,8 @@
   "browserslist": [
     "last 2 versions",
     "> 1%"
-  ]
+  ],
+  "engines": {
+    "node": ">=18.0.0"
+  }
 }
